point_overshoot_fix
Added more failsafes to prevent "overshooting" a point and/or missing a
snap. You're now always guaranteed to snap to a point while passing
through it, and any overshot will get auto-corrected so that it's along
the path towards your next destination. Additionally, the logic for
snapping is now guaranteed to fire if you're at a point and not headed
anywhere, whether you have the joystick tilted or not.A
Also reverted one or two debug changes, e.g. a tweak to the path, the
default path draw width, etc.