pathsloader_rewrite
The PathsLoader class now guarantees that it'll load data back in such
that if you then publish it with PathsWriter, you'll end up with the
same JSON you started with. Basically, this should guarantee that points
and paths are always set up in the same order, and therefore that paths
should always be directed in the same way across loads.