path_deletion
Beginning to add support for on-the-fly editing of the Paths object. To
start, tested out path deletion and manual redraw triggering. Cleaned up
a few method and member variable names to make it clear that they're
really controlling whether the state needs to be refreshed.
Also had to tweak the Paths' draw walk to make sure that it hits every
subsection of a disconnected graph.