path_sculpting
Further tweaks to allow dynamic additiona and deletion of points and
paths. As part of this, also fully removed the Paths.refresh() method,
since it's being fully handled by the dirty test within update().
Subsequently, all methods that manipulate the Paths object in ways which
will require a redraw now flag it as dirty.