path_sculpting2
Cleaned up some of the js directory structure, which unfortunately makes
a few of these diffs messy to see. Basically, continued adding and
refining the user controls for sculpting the Paths object on the fly.
You can now delete paths and points, as well as "merge delete" a point
which will reconnect all points that previously connected to it. Also
added a "float" action which unsnaps the avatar from the paths and lets
it free-roam; releasing will snap it back to the nearest point or path.
Still to come: adding controls to trigger addition of points along
paths, as well as new points stemming off from existing ones (at clean
intervals and angles).