path_sculpting3
Now with support for adding points! You can press X to toggle "Add"
mode.
If used on a path, you can move around and press X again to create a new
point at any of the displayed marks along the path. Marks are displayed
at fixed intervals (currently units of 50).
If used on a point, you get a free-range cursor and a "hint" showing
where your new point+path will be created. Again, points are
automatically adjusted so that they fall onto the grid at intervals of
50.
To make all this work, EditorAvatar also needed some updates to support
the idea of "consuming" button presses. This works by having the editor
able to flag the most recent time when activity was detected on a
particular button. It also exposes methods which will only return true
for a button's pressed/released state if the button has changed since
the last time we recorded an event from it. This allows classes to
listen for a state event and then guarantee that no one else (including
future loops of themself) will trigger off of that particular state
event.