point_angle_to_path_improvements
Added some logic to Point.getPath(angle) so that we won't choose one
path if the angle also closely matches another path. We essentially
enforce a deadzone of 33% between paths; that is to say, for angle a to
match path P1, a must be twice as close to P1 as it is to P2. This means
that any angle within the middle 33% of the space between P1 and P2 will
fail the test and fall into the deadzone.