avatar_path_tilt_handling
Rewrote Path.getPoint() to be *way* more involved.
* It now takes in the (x, y) of your avatar.
* It projects points based on all paths leading away from each of your
endpoints. The new points are projected further away if you're closer to
an endpoint.
* Your angle is gauged against all possible endpoints and projected
points. The minimum angle still wins, but if the minimum is towards a
projected point, your avatar still advances towards the point it was
projected from.
* To complement this, getPoint() now returns an object so that we can
also give back the angle towards the projected point that we're
currently steering towards. This allows us to perform tilt grading based
on that angle instead of the angle towards the endpoint goal (since
those angles may be wildly different, and thus cause our speed to drop
to 0 due to grading even though we're pointing towards a valid
projection).