float_path_snap_fix
For some reason, the FloatAction.findNearbyPath() had a commented-out
call to getBoundedAngleDifference() and was using the absolute value of
an angle subtraction instead. This was causing errors in weird edge
cases where certain coordinates could snap way over to paths that were
nowhere near them.
Also tweaked the background color to something darker, enabled world
bounds collision on the avatar, and cached the Paths object's image for
potential future manipulation.