paths_coordinate_translation
Previously, all coordinate logic was written with the assumption that a
Paths object was rooted at (0, 0) in game coordinates. Now there is a
translation layer which takes place when performing any pointer or
avatar coordinate arithmetic against the Paths. This allows the Paths
object itself to draw all its points relative to its internal (0, 0),
but still be placed at an arbitrary (x, y) within the game area.
To do this, all logic that interacts with the Paths object or its
internals will now do so via translate calls that Paths provides. This
onus is primarily on external input handlers that need to move the
avatar or other cursors relative to the Paths.
Also of minor note, for the AddFromPointIState specifically, it performs
arithmetic mod 50 to find points on a grid. However, this logic
misbehaves if the Paths internal points aren't aligned to units of 50.
To correct this, AddFromPointIState performs a manual offset calculation
to figure out how far the internal points are misaligned from true 50.