load_write_refactor1
Beginning to condense the load/write code. No longer requires
behind-the-scenes instantiation of separate objects for load/write,
making each a single method call. Also reworked the writing in
particular to return a JSONable object instead of a JSON string. This is
a precursor to killing the writer entirely and using the standard
toJSON() override behavior of the objects themselves instead.
Also fixed a small bug where AddIState hadn't been completely updated to
use the new Tier constants yet.