input_event_consume_fixes
Discovered some buggy behavior with the GPad input event consumption and
tried to smooth it out. You now compare against the specific time you're
triggering on (up or down) rather than a best guess. Also had to make
the TitleMenu transition on buttonUp to make sure that event gets fully
consumed so that the PlayLevel doesn't see it.