path_name_and_map_refactoring
Beginning to add names to Path objects. They aren't yet properly set
(since the JSON doesn't have names to pass in), but they exist.
As part of this, removed all methods that create Paths from the Point
and Path classes, so that Tiers can be the defacto factory for Paths
(they already are for points).
All deletion methods on points/paths are also gone now; only the Tier
does that.
Added logic so that the Tier's point and path maps are maintained on the
fly, rather than being rebuilt from scratch during the updateCaches()
loop.