path_name_and_json_format
Finished adding path names, and redid the level JSON so that points list
the path names they connect with, and there's a separate definition for
each path to give its name and endpoints. This should make it easier to
have the JSON define special properties of a path, if desired.