render_vs_update2
Finished moving all the Tier gfx logic over to the render loop instead
of update. This was actually pretty trivial, since (now that IHandlers
tackle input) the tier's update was doing nothing *but* render logic, so
it can move wholesale.
Also fully removed the concept of updateCaches as an externally-called
method (e.g. after load). Adding points/paths to a tier is sufficient to
keep it in a usable state. The only internal recalculations that need to
get done will do so as triggered by the render loop, and should be
transparent to external parties.