regenerated_level_json
Hadn't run level generation since tweaking some of the load/write logic.
We no longer set p0 to (0, 0), but t2 still had that logic which (since
load no longer does a translate, and images can rescale/reposition)
meant that the level was too high and collided with world bounds. The
larger issue with world bounds is already noted in the TODO; eventually,
we'll make them overly large and make sure they scale outwards as needed
to keep a buffer around the image.