broken_path_z-order_fix
Had my boolean logic wrong within Path.setBroken(); this was causing all
paths that had setBroken() called to get the flipped z-order that they
wanted. Since we only called setBroken() on broken paths, the result was
that all paths ended up with the same z-order, and rendered in random
order.