camera_follow_cleanup
Added a deadzone to the camera follow so that it's not quite so
claustrophobic. May need to tune later. For reference, the default
deadzone if you pass Phaser.Camera.FOLLOW_TOPDOWN to the follow() method
is a centered 200x200 rectangle, with margins of x=300 and y=200. So the
custom deadzone (400x200) is allowing you to get slightly closer to the
sides of the stage.