pause_menu_improvements
The pause menu now has an actual UI and functional selection of 3
options: continue, restart, exit. It took a lot to get it working, but
it's straightforward to describe. Everything behaves the way you'd
expect. You can still press start or cancel to leave the menu (i.e.
continue), or press X to select an option that's been highlighted via
the joystick.
TODO: Make the joystick behavior slightly more generic, since it'll
probably be useful for another menu later. In fact, the entire class
should probably be distilled into a generic menu that's derived from a
list of texts + actions.