pause_menu_improvements2
Some more cleanup and gloss. Redid the z-order groups so that there's a
menu group as well as a parent group (of everything *but* menus) for
when you want to affect all in-level groups.
Added a background blur effect when the pause menu is up. It tweens in
and out, too. To get this working properly, the blur had to be applied
to the master group for level objects, but that was doing weird cropping
things where it only affected the area matching the level image. To fix
this, I add an empty bitmap to the level group, with its dimensions
matching the camera view; this seems to help, and shouldn't be too
expensive.
Also went back and figured out how to cause bitmaps to update while
paused (hint: call bitmap.render() during the pauseUpdate), so that I
could properly redraw the selector as desired instead of just scaling
it. Looks cooler now.
Oh, and for some reason the filters weren't being included in Phaser and
the game.load.script technique didn't seem to be working. Rather than
solve it, I've cheated and copied over the filters I want from the
Phaser src into my game project...