imenu_nested_submenus_and_animations
OMG! There is now support for nested submenus. Any menu item can have
subitems added to it. Any menu item can also be defined as a cancel
item, which causes it to be invoked upon a cancel action (if it has an
action) and/or causes its invocation to trigger a retreat back up to the
parent submenu. It's all very complicated, but it seems to work so far.
Also, all menu item transitions are now animated, which requires
tracking a ton of tweens that we have to manually step (thanks,
game.pause).
Of note, also finally renamed a bunch of ButtonMapping constants for
consistency. None of them need _BUTTON at the end, since that's kind of
obvious.