menu_and_color_updates
A bunch more stuff!
* Tiers now have color palettes, and use them during render. Right now
each tier has a temp palette with one color; eventually, these would get
fleshed out to have sets of colors per tier and cross-tier cohesion.
* AvatarGraphicsKey now uses white sprites + tints. The color can be set
via setters on the avatar or the gfx. The actual invocation of this is
now being done based off of the current tier's color palette.
* There is now a settings.edit that controls whether all of the editor
actions are available. Since we're in dev, it's defaulted to true.
* Found (but ultimately left disabled) the setting to turn off
antialiasing. It looks nice on text, but particularly bad on any kind of
rotation sprite.
* Made a ton more changes to the menu so that it can support having its
color changed even after it's been created. This largely involved
methods to allow the chrome's bitmap to be redrawn. Also ultimately
scrapped the "chained tween" hack, since it was somehow misfiring/firing
after the menu reopened; this was leading to it messing with certain
selections' alphas. The chain was really just to artificially delay the
tween rather than slowing it down; it actually looks fine slow, so I've
updated the in/out fades for root items accordingly.