clear_arc_fix
Turns out the issue was that while drawing the portal point, we were
setting the clip for the tier's bitmap and never unsetting it (via
save/restore). This means that future operations on the bitmap were
being treated as clips, I think. Anyways, found an alternative way to do
clears which seems safer(?) I guess.