tier_child_sprites
The gfx for custom points are now attached as child sprites to the tier,
so that they'll scale and move around with it. Note that this does mean
that things like powers are now in z.level instead of z.mg; if this
becomes an issue, consider not having powers scale by design.