point_hold
Points now have a shouldHold() method which tells the avatar whether it
should briefly "stick" to them in passing. This makes it easier for
points along smooth paths to be targeted by the player.
Currently, a point's shouldHold() returns true if that point has a
custom istate for handling input. You could always override shouldHold()
to return a value based on your own criteria, though.