hud_settings1
The "show HUD?" flag is now more granular, supporting always on, always
off, or default fades. Also began to update the pause menu to have
actual contents instead of junk.
TODO: Make sure that the settings submenu for PausedIState is captured
in a generic external class so that it can be easily reused in
TitleMenuIState. Also, start looking at new menu features for things
like args to the option actions, checkboxes and text changes, etc.