ihandle_update_settings_framework
After realizing that ihandlers/istates cache the buttonMap, added an
updateSettings() call hierarchy that makes it easy to propagate new
settings out to all of the ihandlers/istates in a set. Basically, any
ihandler can be told to update, but they'll try to walk up and ask their
outermost wrapping ihandler to actually start the official recursive
update walk, which will then cause each handler and state to do the
updates.