keyburst_gfx
Added some nice gfx when shifting tiers or gaining keys. The tier shift
effect is to help you mentally understand which keys you've now got
access to for spending, and which keys are now in the pot that you can
earn (for spending one tier up).
The effect when a key is added may be temporary, if there end up being
wackier effects for the key particle actually travelling from the level
coords over to the meter.