change_tier_refactor_work
Finally getting around to the change tier logic. Previously this was a
great big mush of code in the level. Moved a lot of the fade logic and
tweens into the tiers themselves. This makes it easier for each tier to
control its own transitions. Also, the tiers now dispatch events as they
begin and end their various fades, and the level now uses these events
to decide when it should hide all of the tier's layers. While it's a bit
jarring, it definitely works!