point_path_fade_listeners
All points and paths now get signed up to listen for the various tier
fade events. This makes it easy to have your subclass override the ones
it cares about.
Consequently, end points now fade the goal rings based on which tier
you're touching.
Also, a disabled end point now stops spinning, so that you can tell from
other tiers whether its active (since the alpha for inactive is the same
as the alpha for "the goal is in another tier").