play_level_state_more_methods_for_override
Moved a lot of the level state's logic into overrideable methods. Also
reworked how the level object loads. It no longer wants a catalog level
(since we may not have one, if we're editing from scratch). This means
that the end istate has to ask for the catalog level from the game state
rather than the level object.