point_break_and_right_angle_merges
Points were considered broken if they had no unbroken paths connected.
Logic is now fixed so that they're only broken if all *1 or more* of
their paths are broken.
Added a check in tier.deletePointAndMerge() so that we only create new
paths that retain right angles.
Also removed the test shutdown() method from the level state; that was
actually a trial for the edit state.