tkey_hover_fix_on_tier_redraw
When the tier redraws, all children get their (x,y) shifted. However, if
a child has already set up a tween on one of those values, the tween
doesn't get updated, and keeps tweening between the old (wrong) values.
To address this in tkey (hopefully that's the only example), added an
empty placeholder sprite object with an internal relative offset of
(0,0), so that we can tween it up and down from a reference point that
doesn't care about shifts in the parent sprite's coordinates.