edit_mode_underway
Starting work in earnest on getting the edit ihandler hierarchy into a
meaningful state. Current plan is to have a single ihandler with all
states registered. Certain key states are listed as modes, and the
handler exposes a nice method to let the modes check for cycle actions
which will change between them. Each of the mode states can then be a
general "check for all the interesting inputs" and fire the appropriate
(sub?)states.