tier_recalculate_fix_again
I was missing the edge case where both dimensions grow, but one does so
positively and one negatively. In this case, we were correctly hitting
the dx/dy < 0 logic which would fix the negative dimension, but we
weren't hitting the dw/dh > 0 which would have fixed the positive
dimension. So, now the latter logic fires always, but factors in whether
dx has run (for dw) or dy has run (for dh).