removed_point_padding
All premade levels no longer have points padded by +5,+5. In updating
the code to expect this, found a bunch of potential logic errors and
just general assumptions which should hopefully be fixed. Note that for
an individual point, its x,y are now explicitly its offset coordinates
from the top left corner of the tier's image. It's gx,gy are its game
coordinates.
And yes, this means that a bunch of logic in Avatar.move() could be
simplified if it used point.g* coords instead of having to translate
through tier.translateInternalPointToGamePoint(point.x, point.y)...