portal_enabled_sync
Portals can now be set up to synchronize their enabledness with their
other end. This basically means that any (future) calls of .setEnabled()
on one end of a portal will also be automatically called against the
other end. Disable sync if you want weird trickiness like a portal which
is one-way until you find a way to power on the way back.