replace_wire_fix_and_contact_switch_infinite_recursion
Fixed so that when first rendering a wire, we use the actual "off" alpha
instead of assuming it's 0 (hint: it's not, in edit mode). Also found
and fixed a possible infinite recursion when stepping on to contact
switches. Not sure how I missed it before.
Basically, now ihandlers will quietly return if you try to (re)activate
the currently active istate.