adjacent_tiers_visible
Beginning to play with making the immediately-adjacent tiers visible as
"ghosts". This requires a lot of retooling to the notions that any
non-current, non-fading tier is fully hidden. I think it also warrants
new signals, since we *also* want those tiers to be able to fade/scale
in and out as they become faintly visible.