wire_state_logic_fixes
We're abusing the base customize class since we're not letting it prompt
for/cancel out of an option. This causes some strife since we have to
auto-fallthrough the update() if we're beginning the action, but somehow
know when we've been canceled back out. The old logic was bugged; this
is better.