add_fails_on_overlap
You can no longer perform an add if it will cause an overlap with
existing coordinates. This logic mainly matters on AddFromPoint.
AddFromPath is literally guaranteed not to introduce conflicts (although
interestingly, it is the usecase which *does* add duplicate entries to
the coords map, since we don't delete the path that gets shortened and
hence its coords are still in the map). AddFromFloat also got the new
logic just in case, even though it seems like it's impossible to trigger
the add at all if there's an overlap, since the "near" tracking only
lets you add if it's not snapping to an existing object. Still better
safe than sorry.