obstacles_collision_and_test_enemies
Added in a collision system that works without completely boning our
path-based movement. Unfortunately, we can't rely on the physics
system's separation behavior to decollide two objects, since it'll
derail the avatar's placement upon a path. So instead we test for
overlap and cut off any further movement. This means that we'll still
overlap a bit, so take that into account. Also, it means that we have to
explicitly check whether the avatar is currently trying to move towards
or away from the obstacle; otherwise, we'd just overlap and then be
stuck forever.