door_and_key_test
Revamped the enemy as a door with gfx and a key to match. Also created
an Obstacle sprite subclass for things like this. This also raises the
question about having a system for picked up/held objects that's generic
enough to not require rewriting.
Oh and also a sytem for writing/loading the JSON for obstacles, and
attaching them to the tier image appropriately. This probably means
having the actual collidable obstacle sprites (which can't be
instantiated without references to the game + tier), and then having
vanilla wrapper objects which can be loaded and have a .draw(tier)
method or similar.