add_obstacles1
Continue to butcher the customize point class hierarchy so that it'll
work for way more things than originally intended. The ultimate, proper
way to handle it would be to just go and refactor the class hierarchy
cleanly, so that points comprise one subclass tree but wires and
obstacles can have their own. That said, this is faster in the meantime.
The base add obstacle class is in; just need to add classes for
individual option types now.