avatar_velocity_kill
The avatar now disables its physics body when it wants to corner/stop
abruptly. This works perfectly except for scenarios where we're
currently colliding with something. In that case, here's the problem: we
only know we're colliding if our body is on, but when we collide we turn
it off to kill velocity, which means we no longer think we're colliding,
which means we allow the body to turn back on with velocity towards our
collider.
Anyways, solved it by leaving the body on in collision scenarios.