tap_through_obstacles_fix
Turns out that we were being too generous in the collision test, and
saying "no collision" if the collision handler was called but it looked
like the avatar wasn't currently moving towards a destination.
Now the collision handler presumes that if it's being called, the avatar
is guilty of collision *unless* it is explicitly moving away from the
obstacle, *or* the obstacle itself says that the avatar shouldn't be
obstructed.
Oh and now the avatar stores its "colliding" state to ensure that
physics *never* gets turned off while it's overlapping an obstruction.