removed_physics_body_velocity
Finally gave up on trying to fix all of the stupid physics issues with
velocity. If you don't disable the body, you get jitter when arriving at
a point. If you *do* disable the body, you get jitter when reenabling it
as you depart a point. Now we just handle the velocity ourselves.