resize_fixes
For the pause menu, we can't fire a delayed event to resize the menu.
So, instead it updates on the next update loop. Also, I think the
tsquare instance that was being used as the context for the painter()
method could actually be dead when you called it from another state
(this is probably an issue for a bunch of painters...). Dead sprites
must have their game object reference nulled out. Long-term, may just
make sense to pass the game object in as another parameter for those who
need it.