death_angle_improvements_and_maskless_death
After thinking about it (and testing a few new cases), decided that the
death angle may as well be from the base of the avatar's hitbox rather
than the mask's center of gravity. This feels better when you very
narrowly clip a death area with your hitbox.
Also added considerations for death gfx when you have no mask.