tier_meter_restore_broadcasts_change_events
When the avatar's restoration updates the shards held in the tier meter,
we now broadcast it (specifically, we broadcast a change for every tier
to make sure everyone gets a chance to update).
This is necessary because things like power points will use this event
to decide whether they're enabled by default.